service: Standardize validation of Repository existence
Majority of RPCs require Repository parameter to be passed.
Some methods doesn't check if it is actually passed or not,
in others the validation error is differs from commonly used.
We try to standardize the approach we use and fix all
inconsistencies. All RPCs that are rely on Repository parameter
check the existence of it on validation step and return
pre-defined error in case it is missing.

Closes: https://gitlab.com/gitlab-org/gitaly/-/issues/3717