cat-file: Switch to using `--batch-command`
Currently both `ObjectReader` and `ObjectInfoReader` use the `--batch`
and `--batch-check` command. With the newly introduced `--batch-command`
in git (v2.36). We could merge the two into one command which does both.
This would remove the need for having two processes running and would
also clean up code around flushing.

In this commit we setup the the base by moving both `ObjectReader` and
`ObjectInfoReader` to using `--batch-command`. This also means that
clients using the `RequestRevision` will now have to explicitly provide
the type of data they're requesting (i.e. info or contents).

In the upcoming commits we'll slowly move all functionality of
`ObjectInfoReader` to `ObjectReader` and deprecate `ObjectInfoReader`.