git: Add gitalyPid option to NewExecCommandFactory
We want all git commands to be tied to a pgid, the gitaly process. This
way, when gitaly gets killed, all git processes can be killed alongside it,
in case there are any zombie processes hanging around.

Add a gitalyPid option to the NewExecCommandFactory() constructor, and
update the callsites with the option.

Changelog: changed