gitpipe: Allow setting LsTree skip function
In 5da3e9a23 (gitpipe: Add LsTree that calls git-ls-tree(1), 2022-06-14)
we've added a skipResult function to the gitpipe.lsTreeConfig, but there
was no way to actually use that feature outside the gitpipe package.

This change adds LsTreeWithSkip() so a user of LsTree() can specify a
skip function, similar to DiffTreeWithSkip().