ref: Return structured error when update error is not ErrAlreadyLocked
We have a bug wherein we still return a successful response with an
error embedded in the response if the Prepare() step fails with an error
but the error is not an ErrAlreadyLocked. Fix the code such that we
return an error instead of a response when there is a non
ErrAlreadyLocked error during the Prepare() call.

Changelog: fixed