Praefect: Stop transactions when quorum impossible
Currently when a secondary node RPC fails the error is ignored and the
transaction continues waiting for additional votes even if there are not
enough outstanding votes to reach quorum. If quorum becomes impossible,
due to the failed node, the transaction hangs until the context gets
cancelled. This is not desirable as ideally once it has been established
that there is not enough outstanding votes to reach the required
threshold specified by the transaction, the transaction should be
stopped. This change adapts the `ErrHandler` function of the secondary
nodes to check if there is enough votes outstanding to still reach
quorum. If quorum is impossible the transaction is stopped and an error
returned. If quorum is still possible the failed node is stored in the
transaction so future subtransactions know to not count a failed nodes
votes towards the total count of outstanding votes. The error is also
not returned since returning an error would result in all RPC streams
being cancelled.