config: Create errors with helpers
We should not use status package to create errors as it could lead
to a double status wrapping with overriding of the initial status
code. Also, it may confuse as it seems like there was a gPRC call
that returned an error. By using own helpers we also can change
implementation without affecting places of usage.